+++ components/datetime/benches/fixtures/mod.rs
@@ -2,35 +2,65 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-pub mod structs;
-
+use icu_datetime::options;
 use icu_datetime::options::DateTimeFormatterOptions;
-use std::fs::File;
-use std::io::BufReader;
+use serde::{Deserialize, Serialize};
 
-#[allow(dead_code)]
-pub fn get_fixture(name: &str) -> std::io::Result<structs::Fixture> {
-    let file = File::open(format!("./benches/fixtures/tests/{name}.json"))?;
-    let reader = BufReader::new(file);
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Fixture(pub Vec<Test>);
 
-    Ok(serde_json::from_reader(reader)?)
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Test {
+    pub setups: Vec<TestInput>,
+    pub values: Vec<String>,
 }
 
-#[allow(dead_code)]
-pub fn get_patterns_fixture() -> std::io::Result<structs::PatternsFixture> {
-    let file = File::open("./benches/fixtures/tests/patterns.json")?;
-    let reader = BufReader::new(file);
-
-    Ok(serde_json::from_reader(reader)?)
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct TestInput {
+    pub locale: String,
+    pub options: TestOptions,
 }
 
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub enum TestOptions {
+    #[serde(rename = "length")]
+    Length(options::length::Bag),
+    #[serde(rename = "components")]
+    #[cfg(feature = "experimental")]
+    Components(options::components::Bag),
+    #[serde(rename = "components")]
+    #[cfg(not(feature = "experimental"))]
+    Components(serde_json::Value),
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct TestOptionsLength {
+    pub date: Option<TestLength>,
+    pub time: Option<TestLength>,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub enum TestLength {
+    #[serde(rename = "short")]
+    Short,
+    #[serde(rename = "medium")]
+    Medium,
+    #[serde(rename = "long")]
+    Long,
+    #[serde(rename = "full")]
+    Full,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct PatternsFixture(pub Vec<String>);
+
 #[allow(dead_code)]
-pub fn get_options(input: &structs::TestOptions) -> Option<DateTimeFormatterOptions> {
+pub fn get_options(input: &TestOptions) -> Option<DateTimeFormatterOptions> {
     match input {
-        structs::TestOptions::Length(bag) => Some((*bag).into()),
+        TestOptions::Length(bag) => Some((*bag).into()),
         #[cfg(feature = "experimental")]
-        structs::TestOptions::Components(bag) => Some((*bag).into()),
+        TestOptions::Components(bag) => Some((*bag).into()),
         #[cfg(not(feature = "experimental"))]
-        structs::TestOptions::Components(_) => None,
+        TestOptions::Components(_) => None,
     }
 }