+++ components/datetime/src/lib.rs
@@ -151,11 +151,13 @@ mod zoned_datetime;
 
 pub use any::{DateFormatter, DateTimeFormatter, ZonedDateTimeFormatter};
 pub use calendar::CldrCalendar;
+#[cfg(feature = "experimental")]
+pub use calendar::InternalCldrCalendar;
 pub use datetime::{TimeFormatter, TypedDateFormatter, TypedDateTimeFormatter};
 pub use error::DateTimeError;
 pub use format::datetime::FormattedDateTime;
 #[cfg(feature = "experimental")]
-pub use format::neo::{FormattedDateTimePattern, TypedDateTimePatternInterpolator};
+pub use format::neo::{FormattedDateTimePattern, TypedDateTimeNames};
 pub use format::time_zone::FormattedTimeZone;
 pub use format::zoned_datetime::FormattedZonedDateTime;
 pub use options::DateTimeFormatterOptions;