+++ provider/baked/properties/src/lib.rs
@@ -3,8 +3,13 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 //! Data for the icu_properties crate
+//!
+//! This data was generated with CLDR version 44.0.0, ICU version release-74-1, and
+//! LSTM segmenter version v0.1.0.
 
 #![no_std]
+// The source is not readable and is massive as HTML.
+#![doc(html_no_source)]
 
 #[cfg(icu4x_custom_data)]
 include!(concat!(core::env!("ICU4X_DATA_DIR"), "/macros.rs"));