+++ provider/datagen/src/transform/icuexport/uprops/bin_cp_set.rs
@@ -38,7 +38,7 @@ macro_rules! expand {
 
                     let mut builder = CodePointInversionListBuilder::new();
                     for (start, end) in &data.ranges {
-                        builder.add_range_u32(&(start..=end));
+                        builder.add_range32(&(start..=end));
                     }
                     let inv_list = builder.build();