+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+- Components
+  - `icu_properties`
+    - Add `Aran` script code (https://github.com/unicode-org/icu4x/pull/4426)
+- FFI:
+    - All languages
+      - Correctly handle invalid UTF8 (https://github.com/unicode-org/icu4x/pull/4353)
+      - Correctly handle chars (https://github.com/unicode-org/icu4x/pull/4349)
+    - JS
+      - Fixed a bug where slice length is computed incorrectly (https://github.com/rust-diplomat/diplomat/pull/372)
+
 ## icu4x 1.4 (Nov 16, 2023)
 
 - General